+++ lc-lib/transports/tcp/common.go
@@ -22,7 +22,6 @@ import (
 	_ "crypto/sha512" // Support for newer SSL signature algorithms
 	"crypto/tls"
 	"errors"
-	"fmt"
 	"net"
 )
 
@@ -74,39 +73,6 @@ type eventsMessage interface {
 	protocolMessage
 	Nonce() *string
 	Events() [][]byte
-}
-
-// parseTLSVersion parses a TLS version string into the tls library value for min/max config
-// We explicitly refuse SSLv3 to mitigate POODLE vulnerability
-func parseTLSVersion(version string, fallback uint16) (uint16, error) {
-	switch version {
-	case "":
-		return fallback, nil
-	case "1.0":
-		return tls.VersionTLS10, nil
-	case "1.1":
-		return tls.VersionTLS11, nil
-	case "1.2":
-		return tls.VersionTLS12, nil
-	case "1.3":
-		return tls.VersionTLS13, nil
-	}
-	return fallback, fmt.Errorf("invalid or unknown TLS version: '%s'", version)
-}
-
-// getTlsVersionAsString returns a string representation of the TLS version
-func getTlsVersionAsString(version uint16) string {
-	switch version {
-	case tls.VersionTLS10:
-		return "TLSv1"
-	case tls.VersionTLS11:
-		return "TLSv1.1"
-	case tls.VersionTLS12:
-		return "TLSv1.2"
-	case tls.VersionTLS13:
-		return "TLSv1.3"
-	}
-	return fmt.Sprintf("Unknown (%d)", version)
 }
 
 func SetClientName(client string) {